+++ src/calculations/stats.rs
@@ -1,4 +1,4 @@
-pub fn energy_from_potential_and_vigilance(potential: u16, vigilance: u16) -> u16 {
+pub fn energy_from_potential_and_vigilance(potential: u32, vigilance: u32) -> u32 {
     let ratio = vigilance as f32 / u16::MAX as f32;
-    (potential as f32 * ratio).floor() as u16
+    (potential as f32 * ratio).floor() as u32
 }
